Consider orientation when calculating InSampleSize during compression
The destination file being overwritten during each iteration of
compression considers the orientation from Exif data, but the calcuation
of InSampleSize does not.

As a result, for phones that don't handle Exif data properly, it
switches the width/height during `calculateInSampleSize` and therefore
would skip that compression step and ultimately fail because it wasn't
able to compress a 3.5MB photo with quality compression using the
`SizeContraint` class alone.

Closes https://github.com/SimpleMobileTools/Simple-SMS-Messenger/issues/511